+++ example/dummy/dummy_test.go
@@ -82,7 +82,7 @@ func TestPersistentDummyInfo(t *testing.T) {
 	}
 	dummy := NewPersistentDummyApplication(dir)
 	InitDummy(dummy)
-	height := uint64(0)
+	height := int64(0)
 
 	resInfo := dummy.Info(types.RequestInfo{})
 	if resInfo.LastBlockHeight != height {
@@ -90,10 +90,10 @@ func TestPersistentDummyInfo(t *testing.T) {
 	}
 
 	// make and apply block
-	height = uint64(1)
+	height = int64(1)
 	hash := []byte("foo")
 	header := &types.Header{
-		Height: uint64(height),
+		Height: int64(height),
 	}
 	dummy.BeginBlock(types.RequestBeginBlock{hash, header})
 	dummy.EndBlock(types.RequestEndBlock{header.Height})
@@ -173,7 +173,7 @@ func TestValSetChanges(t *testing.T) {
 
 func makeApplyBlock(t *testing.T, dummy types.Application, heightInt int, diff []*types.Validator, txs ...[]byte) {
 	// make and apply block
-	height := uint64(heightInt)
+	height := int64(heightInt)
 	hash := []byte("foo")
 	header := &types.Header{
 		Height: height,